feat: create file char-to-textarea.js
