+++ js/main.js
@@ -1,7 +1,6 @@
-import data from './translate.js';
 import { setLang, initKeyboard } from './change-lang.js';
-
 import { setAnimatedSingle, setAnimationDouble } from './animate.js';
+import charToTextarea from './char-to-textarea.js';
 
 window.addEventListener('DOMContentLoaded', () => {
   document.body.insertAdjacentHTML(
@@ -57,35 +56,7 @@ window.addEventListener('DOMContentLoaded', () => {
       }
     } // and shift
 
-    const lang = localStorage.getItem('langKeyBoard');
-    const register = localStorage.getItem('register');
-
-    const key = e.code; // very important
-    let val = document.querySelector('textarea').value; // very important
-    const start = document.querySelector('textarea').selectionStart;
-
-    if (data[lang][register][key]) {
-      if (start >= 0 && start <= val.length) {
-        document.querySelector('textarea').value = val
-          .slice(0, start) + data[lang][register][key] + val.slice(start, val.length);
-        document.querySelector('textarea').selectionStart = start + data[lang][register][key].length;
-        document.querySelector('textarea').selectionEnd = start + data[lang][register][key].length;
-      }
-    } else if (key === 'Backspace') {
-      if (start > 0 && start <= val.length) {
-        val = val.slice(0, start - 1) + val.slice(start, val.length);
-        document.querySelector('textarea').value = val;
-        document.querySelector('textarea').selectionStart = start - 1;
-        document.querySelector('textarea').selectionEnd = start - 1;
-      }
-    } else if (key === 'Delete') {
-      if (start >= 0 && start <= val.length - 1) {
-        val = val.slice(0, start) + val.slice(start + 1, val.length);
-        document.querySelector('textarea').value = val;
-        document.querySelector('textarea').selectionStart = start;
-        document.querySelector('textarea').selectionEnd = start;
-      }
-    }
+    charToTextarea(e); // вводить символы в textarea
     return null;
   }
   document.addEventListener('keydown', keyDownHandler);
@@ -104,11 +75,23 @@ window.addEventListener('DOMContentLoaded', () => {
   }
   document.addEventListener('keyup', keyUpHandler);
 
+  function mouseClickHandler(e) {
+    if (!e.target.classList.contains('output')) {
+      e.preventDefault();
+    }
+
+    // const lang = localStorage.getItem('langKeyBoard');
+    // const register = localStorage.getItem('register');
+    //
+    // const key = e.target.id;
+    // console.log(e.target.id);
+  }
+  document.addEventListener('click', mouseClickHandler);
+
   function mouseDownHandler(e) {
     if (!e.target.classList.contains('output')) {
       e.preventDefault();
     }
-    console.log(e.target.id)
   }
   document.addEventListener('mousedown', mouseDownHandler);
